+++ index.html
@@ -30,8 +30,13 @@
     </button>
     </div>
     
+    <button title="crop" id="cropBtn">
+      <img class="icon" src="assets/crop.ico" />
+    </button>
+
     <div id="imgPanel">
-      <img id="previewImg" draggable="true" src="./assets/addImage.png"/>
+      <canvas id="previewImg" draggable="true" ></canvas>
+      <!-- <img id="previewImg" draggable="true" src="./assets/addImage.png"/> -->
       <div id="mainColorBox">
         <div class="colorBox" id="mainColor1" title="empty"></div>
         <div class="colorBox" id="mainColor2" title="empty"></div>
