+++ renderer/data_handler.js
@@ -4,6 +4,54 @@ var extensionQueue = [];
 
 var i = -1;
 
+var canvas = document.getElementById("previewImg");
+canvas.setAttribute("class", "img-canvas");
+var ctx = canvas.getContext("2d"); 
+
+ipcRenderer.on("openImgCMD", (event, res) => {
+  try {
+    filepath = res;
+    extension = path.extname(filepath).replace(".", "");
+    openImg(filepath, extension);
+  } catch (error) {
+    if (filepath === null) {
+
+      const image = new Image();
+      image.src = "../assets/addImage.png";
+      image.onload = () => {
+           ctx.drawImage(image, 0, 0);
+      }
+      filepath = null;
+      extension = null;
+      buffer = null;
+    }
+  }
+});
+
+ipcRenderer.on("saveImgCMD", (event, res) => {
+  var base64Data = canvas.toDataURL();
+
+  if (!base64Data.includes("base64")) {
+    sharp(filepath).toFile(res.replace(".png", `.${extension}`), (err) => {
+      if (err) {
+        console.log("failed to save");
+      } else {
+        console.log("saved successfully");
+      }
+    });
+  } else {
+    base64Data = base64Data.replace(`data:image/${extension};base64`, "");
+    res = res.replace(".png", `.${extension}`);
+    require("fs").writeFile(res, base64Data, "base64", (err) => {
+      if (err) {
+        console.log("failed to save");
+      } else {
+        console.log("saved successfully");
+      }
+    });
+  }
+});
+
 function openImg(filepath, extension) {
   sharp(filepath).toBuffer((err, buf, info) => {
     updatePreviewImg(buf, info, extension);
@@ -11,8 +59,18 @@ function openImg(filepath, extension) {
 }
 
 function updatePreviewImg(buf, info, extension) {
-  document.getElementById("previewImg").src =
-    `data:image/${extension};base64, ` + buf.toString("base64");
+
+  canvas.width = info.width;
+  canvas.height = info.height;
+
+  const image = new Image();
+  
+  image.src = `data:image/${extension};base64, ` + buf.toString("base64");
+  image.onload = () => {
+       ctx.drawImage(image, 0, 0);
+  }
+
+  console.log(canvas.toDataURL());
 
   updateImgInfoText(info);
   extractMainColors(buf, info);
@@ -58,8 +116,14 @@ function undoPreviewImg() {
     info = infoQueue[i];
     extension = extensionQueue[i];
 
-    document.getElementById("previewImg").src =
-      `data:image/${extension};base64, ` + buffer.toString("base64");
+    canvas.width = infoQueue[i].width;
+    canvas.height = infoQueue[i].height;
+
+    const image = new Image();
+    image.src = `data:image/${extension};base64, ` + bufferQueue[i].toString("base64");
+    image.onload = () => {
+         ctx.drawImage(image, 0, 0)
+    }
 
     updateImgInfoText(info);
     extractMainColors(buffer, info);
@@ -76,8 +140,15 @@ function redoPreviewImg() {
     info = infoQueue[i];
     extension = extensionQueue[i];
 
-    document.getElementById("previewImg").src =
-      `data:image/${extension};base64, ` + buffer.toString("base64");
+    canvas.width = infoQueue[i].width;
+    canvas.height = infoQueue[i].height;
+
+    const image = new Image();
+    image.src = `data:image/${extension};base64, ` + bufferQueue[i].toString("base64");
+    image.onload = () => {
+         ctx.drawImage(image, 0, 0)
+    }
+
     extractMainColors(buffer, info);
     updateImgInfoText(info);
     updateExtension(extension);
@@ -154,3 +225,26 @@ function extractMainColors(buffer, info) {
       }
     });
 }
+
+document
+  .getElementById("previewImg")
+  .addEventListener("drag", function (event) {}, false);
+
+document.getElementById("previewImg").addEventListener(
+  "dragover",
+  function (event) {
+    event.preventDefault();
+  },
+  false
+);
+
+document.getElementById("previewImg").addEventListener(
+  "drop",
+  function (event) {
+    event.preventDefault();
+    filepath = event.dataTransfer.files[0].path;
+    extension = path.extname(filepath).replace(".", "");
+    openImg(filepath, extension);
+  },
+  false
+);
